refactor: Renames resources to loaders
