+++ assets/scripts/components/enso/enso.js
@@ -1,4 +1,4 @@
 // Resource loaders
-export { load } from './utils/resources.js';
+export { load } from './utils/loaders.js';
 // Web component boiler plate base class
 export { default } from './component.js';
