+++ assets/scripts/components/modal/modal.html
@@ -1,9 +1,9 @@
 
-<div id="modal-container" @click="() => this.show = this.static;">
-    <div id="modal-pane" part="modal-pane" @click="e => e.stopPropagation();">
+<div id="modal-container" @click="() => this.show = this.static">
+    <div id="modal-pane" part="modal-pane" @click="e => e.stopPropagation()">
         <div part="modal-header modal-section" class="modal-header modal-section">
             <slot name="modal-header"></slot>
-            <button id="modal-close" part="modal-close" @click="() => this.show = false;">
+            <button id="modal-close" part="modal-close" @click="() => this.show = false">
                 <span aria-hidden="true">&times;</span>
             </button>
         </div>
